+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/MiniAppMessageBridge.kt
@@ -144,6 +144,12 @@ abstract class MiniAppMessageBridge {
             ActionType.SEND_MESSAGE_TO_CONTACT.action -> chatBridge.onSendMessageToContact(
                 callbackObj.id, jsonStr
             )
+            ActionType.SEND_MESSAGE_TO_CONTACT_ID.action -> chatBridge.onSendMessageToContactId(
+                callbackObj.id, jsonStr
+            )
+            ActionType.SEND_MESSAGE_TO_MULTIPLE_CONTACTS.action -> chatBridge.onSendMessageToMultipleContacts(
+                callbackObj.id, jsonStr
+            )
         }
     }