+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/js/chat/ChatBridgeDispatcherSpec.kt
@@ -19,10 +19,21 @@ import org.junit.Before
 import org.junit.Test
 import org.mockito.Mockito
 
+@Suppress("LargeClass")
 class ChatBridgeDispatcherSpec {
     private lateinit var miniAppBridge: MiniAppMessageBridge
-    private val sendMessageCallbackObj = CallbackObj(
+    private val singleChatCallbackObj = CallbackObj(
         action = ActionType.SEND_MESSAGE_TO_CONTACT.action,
+        param = null,
+        id = TEST_CALLBACK_ID
+    )
+    private val multipleChatCallbackObj = CallbackObj(
+        action = ActionType.SEND_MESSAGE_TO_MULTIPLE_CONTACTS.action,
+        param = null,
+        id = TEST_CALLBACK_ID
+    )
+    private val specificIdCallbackObj = CallbackObj(
+        action = ActionType.SEND_MESSAGE_TO_CONTACT_ID.action,
         param = null,
         id = TEST_CALLBACK_ID
     )
@@ -45,7 +56,11 @@ class ChatBridgeDispatcherSpec {
     }
 
     private fun createChatMessageBridgeDispatcher(
-        canSendMessage: Boolean
+        canSendMessage: Boolean,
+        isCancelSingleOp: Boolean = false,
+        isCancelContactIdOp: Boolean = false,
+        isDifferentContactIdOp: Boolean = false,
+        isCancelMultipleOp: Boolean = false
     ): ChatBridgeDispatcher = object : ChatBridgeDispatcher {
 
         override fun sendMessageToContact(
@@ -53,14 +68,41 @@ class ChatBridgeDispatcherSpec {
             onSuccess: (contactId: String?) -> Unit,
             onError: (message: String) -> Unit
         ) {
-            if (canSendMessage)
-                onSuccess.invoke(TEST_CONTACT.id)
-            else
-                onError.invoke(TEST_ERROR_MSG)
+            when {
+                isCancelSingleOp -> onSuccess.invoke(null)
+                canSendMessage -> onSuccess.invoke(TEST_CONTACT.id)
+                else -> onError.invoke(TEST_ERROR_MSG)
+            }
+        }
+
+        override fun sendMessageToContactId(
+            contactId: String,
+            message: MessageToContact,
+            onSuccess: (contactId: String?) -> Unit,
+            onError: (message: String) -> Unit
+        ) {
+            when {
+                isCancelContactIdOp -> onSuccess.invoke(null)
+                canSendMessage -> onSuccess.invoke(TEST_CONTACT.id)
+                isDifferentContactIdOp -> onSuccess.invoke("different_contact_id")
+                else -> onError.invoke(TEST_ERROR_MSG)
+            }
+        }
+
+        override fun sendMessageToMultipleContacts(
+            message: MessageToContact,
+            onSuccess: (contactIds: List<String>?) -> Unit,
+            onError: (message: String) -> Unit
+        ) {
+            when {
+                isCancelMultipleOp -> onSuccess.invoke(emptyList())
+                canSendMessage -> onSuccess.invoke(listOf(TEST_CONTACT.id))
+                else -> onError.invoke(TEST_ERROR_MSG)
+            }
         }
     }
 
-    private val messageToContact = MessageToContact("", "", "", "", "")
+    private val messageToContact = MessageToContact("", "", "", "")
 
     private fun createChatBridge(
         chatBridgeDispatcher: ChatBridgeDispatcher,
@@ -87,13 +129,61 @@ class ChatBridgeDispatcherSpec {
         )
     )
 
+    private val multipleMessageJsonStr = Gson().toJson(
+        CallbackObj(
+            action = ActionType.SEND_MESSAGE_TO_MULTIPLE_CONTACTS.action,
+            param = SendContactCallbackObj.MessageParam(messageToContact),
+            id = TEST_CALLBACK_ID
+        )
+    )
+
+    private val specificMessageJsonStr = Gson().toJson(
+        CallbackObj(
+            action = ActionType.SEND_MESSAGE_TO_CONTACT_ID.action,
+            param = SendContactIdCallbackObj.MessageParamId("contactId", messageToContact),
+            id = TEST_CALLBACK_ID
+        )
+    )
+
+    private val chatBridgeOnSuccess =
+        Mockito.spy(
+            createChatBridge(
+                Mockito.spy(
+                    createChatMessageBridgeDispatcher(true)
+                ), true
+            )
+        )
+
+    /** region: doesn't implementation area */
+    @Test
+    fun `postError should be called when hostapp doesn't implement the chat dispatcher for single chat`() {
+        val chatMessageBridgeDispatcher =
+            Mockito.spy(createChatMessageBridgeDispatcher(false))
+        val chatBridge = Mockito.spy(createChatBridge(chatMessageBridgeDispatcher, false))
+        val errMsg = "The `ChatBridgeDispatcher` has not been implemented by the Host App."
+
+        chatBridge.onSendMessageToContact(singleChatCallbackObj.id, sendingMessageJsonStr)
+        verify(bridgeExecutor).postError(singleChatCallbackObj.id, errMsg)
+    }
+
+    @Test
+    fun `postError should be called when hostapp doesn't implement the chat dispatcher for multiple chat`() {
+        val chatMessageBridgeDispatcher =
+            Mockito.spy(createChatMessageBridgeDispatcher(false))
+        val chatBridge = Mockito.spy(createChatBridge(chatMessageBridgeDispatcher, false))
+        val errMsg = "The `ChatBridgeDispatcher` has not been implemented by the Host App."
+
+        chatBridge.onSendMessageToMultipleContacts(multipleChatCallbackObj.id, multipleMessageJsonStr)
+        verify(bridgeExecutor).postError(multipleChatCallbackObj.id, errMsg)
+    }
+
     @Test
     fun `postError should be called when hostapp doesn't implement the chat dispatcher`() {
         val chatMessageBridgeDispatcher = Mockito.spy(createChatMessageBridgeDispatcher(false))
         val chatBridge = Mockito.spy(createChatBridge(chatMessageBridgeDispatcher, false))
 
-        chatBridge.onSendMessageToContact(sendMessageCallbackObj.id, sendingMessageJsonStr)
-        verify(bridgeExecutor).postError(sendMessageCallbackObj.id, ErrorBridgeMessage.NO_IMPL)
+        chatBridge.onSendMessageToContact(specificIdCallbackObj.id, sendingMessageJsonStr)
+        verify(bridgeExecutor).postError(specificIdCallbackObj.id, ErrorBridgeMessage.NO_IMPL)
     }
 
     @Test
@@ -102,16 +192,91 @@ class ChatBridgeDispatcherSpec {
         val chatBridge = Mockito.spy(createChatBridge(chatMessageBridgeDispatcher, true))
         val errMsg = "$ERR_SEND_MESSAGE $TEST_ERROR_MSG"
 
-        chatBridge.onSendMessageToContact(sendMessageCallbackObj.id, sendingMessageJsonStr)
-        verify(bridgeExecutor).postError(sendMessageCallbackObj.id, errMsg)
+        chatBridge.onSendMessageToContact(singleChatCallbackObj.id, sendingMessageJsonStr)
+        verify(bridgeExecutor).postError(singleChatCallbackObj.id, errMsg)
     }
 
     @Test
-    fun `postValue should be called when hostapp can send message`() {
-        val chatMessageBridgeDispatcher = Mockito.spy(createChatMessageBridgeDispatcher(true))
+    fun `postError should be called when hostapp can't send message to multiple contacts`() {
+        val chatMessageBridgeDispatcher = Mockito.spy(createChatMessageBridgeDispatcher(false))
         val chatBridge = Mockito.spy(createChatBridge(chatMessageBridgeDispatcher, true))
+        val errMsg = "$ERR_SEND_MESSAGE $TEST_ERROR_MSG"
 
-        chatBridge.onSendMessageToContact(sendMessageCallbackObj.id, sendingMessageJsonStr)
-        verify(bridgeExecutor).postValue(sendMessageCallbackObj.id, TEST_CONTACT.id)
+        chatBridge.onSendMessageToContact(multipleChatCallbackObj.id, multipleMessageJsonStr)
+        verify(bridgeExecutor).postError(multipleChatCallbackObj.id, errMsg)
+    }
+
+    @Test
+    fun `postError should be called when hostapp can't send message to a contact id`() {
+        val chatMessageBridgeDispatcher =
+            Mockito.spy(createChatMessageBridgeDispatcher(false))
+        val chatBridge = Mockito.spy(createChatBridge(chatMessageBridgeDispatcher, true))
+        val errMsg = "$ERR_SEND_MESSAGE $TEST_ERROR_MSG"
+
+        chatBridge.onSendMessageToContact(specificIdCallbackObj.id, specificMessageJsonStr)
+        verify(bridgeExecutor).postError(specificIdCallbackObj.id, errMsg)
+    }
+
+    /** end region */
+
+    /** region: onSuccess with value */
+    @Test
+    fun `postValue should be called when hostapp can send message to single contact`() {
+        chatBridgeOnSuccess.onSendMessageToContact(singleChatCallbackObj.id, sendingMessageJsonStr)
+        verify(bridgeExecutor).postValue(singleChatCallbackObj.id, TEST_CONTACT.id)
+    }
+
+    @Test
+    fun `postValue should be called when hostapp can send message to specific contact id`() {
+        chatBridgeOnSuccess.onSendMessageToContact(specificIdCallbackObj.id, specificMessageJsonStr)
+        verify(bridgeExecutor).postValue(specificIdCallbackObj.id, TEST_CONTACT.id)
+    }
+
+    @Test
+    fun `postValue should be called when hostapp can send message to multiple contacts`() {
+        chatBridgeOnSuccess.onSendMessageToMultipleContacts(
+            multipleChatCallbackObj.id, multipleMessageJsonStr
+        )
+        verify(bridgeExecutor).postValue(
+            multipleChatCallbackObj.id,
+            Gson().toJson(listOf(TEST_CONTACT.id)).toString()
+        )
+    }
+
+    /** end region */
+
+    /** region: onSuccess with null when cancellation */
+    @Test
+    fun `postValue should be called when hostapp wants to cancel sending message to single contact`() {
+        val dispatcher = Mockito.spy(createChatMessageBridgeDispatcher(false, true, false, false, false))
+        val chatBridge = Mockito.spy(createChatBridge(dispatcher, true))
+        chatBridge.onSendMessageToContact(singleChatCallbackObj.id, sendingMessageJsonStr)
+        verify(bridgeExecutor).postValue(singleChatCallbackObj.id, "null")
+    }
+
+    @Test
+    fun `postValue should be called when hostapp wants to cancel sending message to specific contact id`() {
+        val dispatcher = Mockito.spy(createChatMessageBridgeDispatcher(false, false, true, false, false))
+        val chatBridge = Mockito.spy(createChatBridge(dispatcher, true))
+        chatBridge.onSendMessageToContactId(specificIdCallbackObj.id, specificMessageJsonStr)
+        verify(bridgeExecutor).postValue(specificIdCallbackObj.id, "null")
+    }
+
+    @Test
+    fun `postValue should be called when hostapp wants to cancel sending message to multiple contacts`() {
+        val dispatcher = Mockito.spy(createChatMessageBridgeDispatcher(false, false, false, false, true))
+        val chatBridge = Mockito.spy(createChatBridge(dispatcher, true))
+        chatBridge.onSendMessageToMultipleContacts(multipleChatCallbackObj.id, multipleMessageJsonStr)
+        verify(bridgeExecutor).postValue(multipleChatCallbackObj.id, "null")
+    }
+
+    /** end region */
+
+    @Test
+    fun `postValue should be called with null when hostapp wants to send a different specific contact id`() {
+        val dispatcher = Mockito.spy(createChatMessageBridgeDispatcher(false, false, false, true, false))
+        val chatBridge = Mockito.spy(createChatBridge(dispatcher, true))
+        chatBridge.onSendMessageToContactId(specificIdCallbackObj.id, specificMessageJsonStr)
+        verify(bridgeExecutor).postValue(specificIdCallbackObj.id, "null")
     }
 }